refactor: make files in `lib` kebab-case
