+++ lib/event-colors.ts
