+++ lib/extract-portraits.tsx
